requirements.txt
asgiref==3.6.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
cryptography==41.0.1
defusedxml==0.7.1
Django==4.2.1
django-admin-honeypot-updated-2021==1.2.0
django-admin-thumbnails==0.2.8
django-allauth==0.54.0
django-auto-logout==0.5.1
django-ipware==5.0.0
django-js-asset==2.0.0
django-mptt==0.14.0
django-otp==1.2.1
django-password-validators==1.7.1
django-phonenumber-field==7.1.0
django-ranged-response==0.2.0
django-recaptcha==3.0.0
emoji==2.5.0
idna==3.4
install==1.3.5
oauthlib==3.2.2
paypal==1.2.5
phonenumberslite==8.13.16
Pillow==9.5.0
psycopg2==2.9.7
psycopg2-binary==2.9.7
pycparser==2.21
PyJWT==2.7.0
pypng==0.20220715.0
python-decouple==3.8
python3-openid==3.2.0
pytz==2023.3
qrcode==7.4.2
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.2
